Fixed: changed the start_mark and end_mark within cmyth_commbreak to be long long to match the transmission encoding of the marks in QUERY_COMMBREAK via the Myth Protocol and also the result set conversion done through the mysql interface. (thanks davilla)
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/linuxport/XBMC@22197 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/lib/libcmyth/cmyth.h
@@ -122,9 +122,9 @@ struct cmyth_file;
 typedef struct cmyth_file *cmyth_file_t;
 
 struct cmyth_commbreak {
-        long start_mark;
+        long long start_mark;
         long long start_offset;
-        long end_mark;
+        long long end_mark;
         long long end_offset;
 };
 typedef struct cmyth_commbreak *cmyth_commbreak_t;
